# Conflicts:
#	server/src/main/java/org/apache/iotdb/db/conf/IoTDBConfig.java
#	server/src/main/java/org/apache/iotdb/db/engine/StorageEngine.java

+++ tsfile/src/main/java/org/apache/iotdb/tsfile/write/TsFileWriter.java
@@ -273,7 +273,7 @@ public class TsFileWriter implements AutoCloseable {
    * @return total memory size used
    */
   private long calculateMemSizeForAllGroup() {
-    int memTotalSize = 0;
+    long memTotalSize = 0;
     for (IChunkGroupWriter group : groupWriters.values()) {
       memTotalSize += group.updateMaxGroupMemSize();
     }
